+++ src/models/Board.ts
@@ -4,7 +4,7 @@ import { getRandomItem } from "../getRandomItem";
 import * as WordMod from "./Word";
 import { Word } from './Word';
 import { Letter } from "./Letter";
-import { SuccessStatus } from "../SuccessStatus";
+import { SuccessStatus } from "./SuccessStatus";
 
 export interface Board {
   words: Word[];
